requirements.txt
python-binance
python-dotenv==1.0.1
pandas>=1.5.0
numpy>=1.23.5
pandas-ta>=0.3.14b
ccxt>=4.0.0
schedule>=1.2.0
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.15.0
scikit-learn>=1.3.0
ta-lib>=0.4.25
yfinance>=0.2.0